Create helper functions to manipulate the lists in order to generate the Identicon grid.
pick_color receives the image hash by pattern matching it with the struct and returns
the list with the tree colors (the three first items).

build_grid uses the same pattern match but tu separate the items into small three sized
chunks (lists) of data to redirect them to mirror_row, mapping all of the items with the
function.

mirror_row takes one row and pattern match the first and second elements, to then append
them to the row itself, "mirrorring" the results.