Add new functions to generate identicon image
filter_odd_squares will check the list for the odd numbers
in order to determine if they will be colored or not. If its odd,
it will be blank, otherwise, it will be colored.

build_pixel_map creates top_left, bottom_right data structures to
generate a 250px x 250px rectangle

draw_image uses the :egd module to generate images, drawing it
based on the size of the rectangle and coloring it.

save_image saves the image to the drive